pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xa-transactions"
version = "0.1.0"
description = "MySQL XA (2-phase commit) transaction coordination library for parallel workers"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mysql", "xa", "distributed-transactions", "2pc", "celery"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mysqlclient>=2.1.0; platform_system != 'Windows'",
    "PyMySQL>=1.0.0",
]

[project.optional-dependencies]
celery = [
    "celery>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/xa-transactions"
Documentation = "https://github.com/yourusername/xa-transactions#readme"
Repository = "https://github.com/yourusername/xa-transactions"

[tool.setuptools.packages.find]
where = ["."]
include = ["xa_transactions*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
